Report disabled TestTemplate methods to TestWatchers
Fixes #2355.
