+++ junit-jupiter-engine/src/main/java/org/junit/jupiter/engine/descriptor/MethodBasedTestDescriptor.java
@@ -15,13 +15,23 @@ import static org.junit.jupiter.engine.descriptor.DisplayNameUtils.determineDisp
 
 import java.lang.reflect.Method;
 import java.util.LinkedHashSet;
+import java.util.List;
 import java.util.Optional;
 import java.util.Set;
+import java.util.function.Consumer;
 
 import org.apiguardian.api.API;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.api.extension.TestWatcher;
 import org.junit.jupiter.engine.config.JupiterConfiguration;
+import org.junit.jupiter.engine.execution.JupiterEngineExecutionContext;
+import org.junit.jupiter.engine.extension.ExtensionRegistry;
+import org.junit.platform.commons.logging.Logger;
+import org.junit.platform.commons.logging.LoggerFactory;
 import org.junit.platform.commons.util.ClassUtils;
 import org.junit.platform.commons.util.Preconditions;
+import org.junit.platform.commons.util.ReflectionUtils;
+import org.junit.platform.commons.util.UnrecoverableExceptions;
 import org.junit.platform.engine.TestDescriptor;
 import org.junit.platform.engine.TestTag;
 import org.junit.platform.engine.UniqueId;
@@ -36,8 +46,9 @@ import org.junit.platform.engine.support.hierarchical.ExclusiveResource;
 @API(status = INTERNAL, since = "5.0")
 public abstract class MethodBasedTestDescriptor extends JupiterTestDescriptor {
 
-	private final Class<?> testClass;
+	private static final Logger logger = LoggerFactory.getLogger(MethodBasedTestDescriptor.class);
 
+	private final Class<?> testClass;
 	private final Method testMethod;
 
 	/**
@@ -90,6 +101,49 @@ public abstract class MethodBasedTestDescriptor extends JupiterTestDescriptor {
 	public String getLegacyReportingName() {
 		return String.format("%s(%s)", testMethod.getName(),
 			ClassUtils.nullSafeToString(Class::getSimpleName, testMethod.getParameterTypes()));
+	}
+
+	/**
+	 * Invoke {@link TestWatcher#testDisabled(ExtensionContext, Optional)} on each
+	 * registered {@link TestWatcher}, in registration order.
+	 *
+	 * @since 5.4
+	 */
+	@Override
+	public void nodeSkipped(JupiterEngineExecutionContext context, TestDescriptor descriptor, SkipResult result) {
+		if (context != null) {
+			invokeTestWatchers(context, false,
+				watcher -> watcher.testDisabled(context.getExtensionContext(), result.getReason()));
+		}
+	}
+
+	/**
+	 * @since 5.4
+	 */
+	protected void invokeTestWatchers(JupiterEngineExecutionContext context, boolean reverseOrder,
+			Consumer<TestWatcher> callback) {
+
+		ExtensionRegistry registry = context.getExtensionRegistry();
+
+		List<TestWatcher> watchers = reverseOrder //
+				? registry.getReversedExtensions(TestWatcher.class)
+				: registry.getExtensions(TestWatcher.class);
+
+		watchers.forEach(watcher -> {
+			try {
+				callback.accept(watcher);
+			}
+			catch (Throwable throwable) {
+				UnrecoverableExceptions.rethrowIfUnrecoverable(throwable);
+				ExtensionContext extensionContext = context.getExtensionContext();
+				logger.warn(throwable,
+					() -> String.format("Failed to invoke TestWatcher [%s] for method [%s] with display name [%s]",
+						watcher.getClass().getName(),
+						ReflectionUtils.getFullyQualifiedMethodName(extensionContext.getRequiredTestClass(),
+							extensionContext.getRequiredTestMethod()),
+						getDisplayName()));
+			}
+		});
 	}
 
 }
