+++ junit-jupiter-engine/src/main/java/org/junit/jupiter/engine/descriptor/TestMethodTestDescriptor.java
@@ -15,9 +15,6 @@ import static org.junit.jupiter.engine.descriptor.ExtensionUtils.populateNewExte
 import static org.junit.jupiter.engine.support.JupiterThrowableCollectorFactory.createThrowableCollector;
 
 import java.lang.reflect.Method;
-import java.util.List;
-import java.util.Optional;
-import java.util.function.Consumer;
 
 import org.apiguardian.api.API;
 import org.junit.jupiter.api.TestInstance.Lifecycle;
@@ -41,9 +38,6 @@ import org.junit.jupiter.engine.execution.ExecutableInvoker.ReflectiveIntercepto
 import org.junit.jupiter.engine.execution.JupiterEngineExecutionContext;
 import org.junit.jupiter.engine.extension.ExtensionRegistry;
 import org.junit.jupiter.engine.extension.MutableExtensionRegistry;
-import org.junit.platform.commons.logging.Logger;
-import org.junit.platform.commons.logging.LoggerFactory;
-import org.junit.platform.commons.util.ReflectionUtils;
 import org.junit.platform.commons.util.UnrecoverableExceptions;
 import org.junit.platform.engine.TestDescriptor;
 import org.junit.platform.engine.TestExecutionResult;
@@ -73,7 +67,6 @@ public class TestMethodTestDescriptor extends MethodBasedTestDescriptor {
 
 	public static final String SEGMENT_TYPE = "method";
 	private static final ExecutableInvoker executableInvoker = new ExecutableInvoker();
-	private static final Logger logger = LoggerFactory.getLogger(TestMethodTestDescriptor.class);
 	private static final ReflectiveInterceptorCall<Method, Void> defaultInterceptorCall = ReflectiveInterceptorCall.ofVoidMethod(
 		InvocationInterceptor::interceptTestMethod);
 
@@ -278,20 +271,6 @@ public class TestMethodTestDescriptor extends MethodBasedTestDescriptor {
 	}
 
 	/**
-	 * Invoke {@link TestWatcher#testDisabled(ExtensionContext, Optional)} on each
-	 * registered {@link TestWatcher}, in registration order.
-	 *
-	 * @since 5.4
-	 */
-	@Override
-	public void nodeSkipped(JupiterEngineExecutionContext context, TestDescriptor descriptor, SkipResult result) {
-		if (context != null) {
-			invokeTestWatchers(context, false,
-				watcher -> watcher.testDisabled(context.getExtensionContext(), result.getReason()));
-		}
-	}
-
-	/**
 	 * Invoke {@link TestWatcher#testSuccessful testSuccessful()},
 	 * {@link TestWatcher#testAborted testAborted()}, or
 	 * {@link TestWatcher#testFailed testFailed()} on each
@@ -322,35 +301,6 @@ public class TestMethodTestDescriptor extends MethodBasedTestDescriptor {
 				}
 			});
 		}
-	}
-
-	/**
-	 * @since 5.4
-	 */
-	private void invokeTestWatchers(JupiterEngineExecutionContext context, boolean reverseOrder,
-			Consumer<TestWatcher> callback) {
-
-		ExtensionRegistry registry = context.getExtensionRegistry();
-
-		List<TestWatcher> watchers = reverseOrder //
-				? registry.getReversedExtensions(TestWatcher.class)
-				: registry.getExtensions(TestWatcher.class);
-
-		watchers.forEach(watcher -> {
-			try {
-				callback.accept(watcher);
-			}
-			catch (Throwable throwable) {
-				UnrecoverableExceptions.rethrowIfUnrecoverable(throwable);
-				ExtensionContext extensionContext = context.getExtensionContext();
-				logger.warn(throwable,
-					() -> String.format("Failed to invoke TestWatcher [%s] for method [%s] with display name [%s]",
-						watcher.getClass().getName(),
-						ReflectionUtils.getFullyQualifiedMethodName(extensionContext.getRequiredTestClass(),
-							extensionContext.getRequiredTestMethod()),
-						getDisplayName()));
-			}
-		});
 	}
 
 	/**
